Bind post_task lazily in TaskEnhancements#after.
Consider you have two files:

    # a.rake

    task :a do
    end
    after :a, :b

    # b.rake

    task :b do
    end

If `a.rake` is loaded before `b.rake`, it will fail, complaining that
"Don't know how to build task b". This is because TaskEnhancements#after
tries to look up task `b` before it's been defined.

However, the task is not actually needed until execution gets inside the
block that's passed to Rake::Task#enhance, at which point all tasks
should have been loaded and it's legitimate to complain about a task not
being recognised.

Havign this change allows app authors to write something like

    Dir.glob('lib/capistrano/tasks/*.rake').each { |r| import r }

in their Capfile and not explicitly define load order. This cannot be
solved by calling `require` at the top of `a.rake` because `require`
throws a `LoadError` on `.rake` files. It also cannot be solved by
calling `import` at the top of `a.rake`, because Rake's `import` does
not load the referenced file immediately; it puts it in a queue and
continues processing the current file.